lattice_model.py split into multiple files
